Fixed animated sprites only being able to play a single animation when there are multiple instances of the same sprite in the scene
Added prompt when user changes scene with unsaved changes
Moved Scripts into their own folder
Fixed renaming in content explorer
Moved Components.h out of components folder

+++ Engine/src/Renderer/Tileset.cpp
@@ -122,7 +122,7 @@ bool Tileset::SaveAs(const std::filesystem::path& filepath) const
 		if (m_Tiles[i].GetProbability() != 1.0 || m_Tiles[i].GetCollisionShape() != Tile::CollisionShape::None)
 		{
 			tinyxml2::XMLElement* pTile = pRoot->InsertNewChildElement("Tile");
-			pTile->SetAttribute("Id", i);
+			pTile->SetAttribute("Id", (int64_t)i);
 			pTile->SetAttribute("Probability", m_Tiles[i].GetProbability());
 			pTile->SetAttribute("Shape", (int)m_Tiles[i].GetCollisionShape());
 		}
